+++ certbot/constants.py
@@ -109,6 +109,7 @@ CLI_DEFAULTS = dict(
     dns_digitalocean=False,
     dns_dnsimple=False,
     dns_dnsmadeeasy=False,
+    dns_dreamhost=False,
     dns_gehirn=False,
     dns_google=False,
     dns_linode=False,