+++ certbot/plugins/disco.py
@@ -29,6 +29,7 @@ class PluginEntryPoint(object):
         "certbot-dns-digitalocean",
         "certbot-dns-dnsimple",
         "certbot-dns-dnsmadeeasy",
+        "certbot-dns-dreamhost",
         "certbot-dns-gehirn",
         "certbot-dns-google",
         "certbot-dns-linode",